refactoring: test package reorg
