+++ core/src/test/java/module-info.java
@@ -36,4 +36,5 @@ open module io.questdb.test {
 
     exports io.questdb.test;
     exports io.questdb.test.cairo;
+    exports io.questdb.test.std.filewatch;
 }
